+++ settings.php
@@ -244,6 +244,20 @@ if ($ADMIN->fulltree) {
         $setting->set_updatedcallback('theme_reset_all_caches');
     $page->add($setting);
 
+    // Settings title to group footer related settings together with a common heading. We don't want a description here.
+    $setting = new admin_setting_heading('theme_boost_campus/navdrawerheading',
+        get_string('navdrawerheadingsetting', 'theme_boost_campus', null, true), null);
+    $page->add($setting);
+
+    // Set navdrwaer to full width on small screens when expanded.
+    $setting = new admin_setting_configcheckbox('theme_boost_campus/nawdrawerfullwidth',
+        get_string('nawdrawerfullwidthsetting', 'theme_boost_campus', null, true),
+        get_string('nawdrawerfullwidthsettings_desc', 'theme_boost_campus', null, true), 'no', 'yes', 'no' ); // Overriding default values
+        // yes = 1 and no = 0 because of the use of empty() in theme_boost_campus_get_pre_scss() (lib.php). Default 0 value would
+        // not write the variable to scss that could cause the scss to crash if used in that file. See MDL-58376.
+        $setting->set_updatedcallback('theme_reset_all_caches');
+    $page->add($setting);
+
     // Add tab to settings page.
     $settings->add($page);